// src/models/user.ts
export class Friend {
  username: string = '';
  fullname: string = '';
  avatar: string = '';
  email: string = '';
  createdAt:string = '';
  updatedAt:string = '';
}

export class User {
  username: string = '';
  password: string = '';
  fullname: string = '';
  avatar: string = '';
  email: string = '';
}